Adjusted image path to allow create cabin without image

+++ src/services/apiCabins.js
@@ -11,29 +11,48 @@ export async function getCabins() {
   return data;
 }
 
-export async function createCabin(newCabin) {
-  const imageName = `${Math.random()}-${newCabin.image.name}`.replace("/", "");
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }]);
+  let imageName = null;
+  let imagePath = null;
+
+  if (newCabin.image) {
+    imageName = `${Math.random()}-${newCabin.image.name}`.replace("/", "");
+
+    imagePath = hasImagePath
+      ? newCabin.image
+      : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  }
+
+  let query = supabase.from("cabins");
+
+  if (!id) {
+    query = query.insert([{ ...newCabin, image: imagePath }]);
+  }
+
+  if (id) {
+    query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+  }
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.log(error);
     throw new Error("Could not create a cabin");
   }
 
-  const { error: storageError } = await supabase.storage
-    .from("cabin-images")
-    .upload(imageName, newCabin.image);
+  if (!hasImagePath && newCabin.image) {
+    const { error: storageError } = await supabase.storage
+      .from("cabin-images")
+      .upload(imageName, newCabin.image);
 
-  if (storageError) {
-    await supabase.from("cabins").delete().eq("id", data.id);
-    console.log(storageError);
-    throw new Error("Cabin image could not be uploaded");
+    if (storageError) {
+      await supabase.from("cabins").delete().eq("id", data.id);
+      console.log(storageError);
+      throw new Error("Cabin image could not be uploaded");
+    }
   }
-
   return data;
 }
